Remove executable policy from DTO packages
Keeps public DTO roots limited to contracts and narrow codecs while relocating GitHub and Secrets policy to their owning implementations.

Fixes ENG-1147
