+++ .changeset/quiet-otters-migrate.md
@@ -1,0 +1,8 @@
+---
+"@shipfox/api-integration-core-dto": major
+"@shipfox/api-integration-github": minor
+"@shipfox/api-integration-github-dto": major
+"@shipfox/api-secrets-dto": major
+---
+
+Removes executable policy and test fixtures from public API DTO roots.
