+++ apps/docs/package.json
@@ -37,6 +37,7 @@
   "devDependencies": {
     "@shipfox/api-agent-dto": "workspace:*",
     "@shipfox/api-integration-core-dto": "workspace:*",
+    "@shipfox/api-integration-github": "workspace:*",
     "@shipfox/api-integration-github-dto": "workspace:*",
     "@shipfox/api-integration-sentry-dto": "workspace:*",
     "@shipfox/api-integration-webhook-dto": "workspace:*",
