+++ libs/api/integration/github-dto/src/index.ts
@@ -1,3 +1,2 @@
-export * from './agent-tools/catalog.js';
 export * from './event-catalog.js';
 export * from './schemas/index.js';
