+++ libs/api/secrets-dto/src/schemas/warnings.test.ts
@@ -3,18 +3,9 @@ import {
   SENSITIVE_VARIABLE_NAME_WARNING,
   SHORT_SECRET_VALUE_WARNING,
   secretWriteWarningSchema,
-  shouldWarnSensitiveVariableName,
-  shouldWarnShortSecretValue,
 } from './warnings.js';
 
 describe('secret write warnings', () => {
-  it('classifies short secret values and sensitive variable names', () => {
-    expect(shouldWarnShortSecretValue('short', 12)).toBe(true);
-    expect(shouldWarnShortSecretValue('long-enough-value', 12)).toBe(false);
-    expect(shouldWarnSensitiveVariableName('API_TOKEN')).toBe(true);
-    expect(shouldWarnSensitiveVariableName('REGION')).toBe(false);
-  });
-
   it('validates warning payloads', () => {
     expect(
       secretWriteWarningSchema.safeParse({code: SHORT_SECRET_VALUE_WARNING, key: 'TOKEN'}).success,
