+++ libs/api/secrets-dto/src/schemas/warnings.ts
@@ -1,5 +1,5 @@
 import {z} from 'zod';
-import {isSensitiveSecretName, isShortSecretValue, secretKeySchema} from './identifiers.js';
+import {secretKeySchema} from './identifiers.js';
 
 export const SHORT_SECRET_VALUE_WARNING = 'short-secret-value';
 export const SENSITIVE_VARIABLE_NAME_WARNING = 'sensitive-variable-name';
@@ -10,11 +10,3 @@ export const secretWriteWarningSchema = z.object({
 });
 
 export type SecretWriteWarningDto = z.infer<typeof secretWriteWarningSchema>;
-
-export function shouldWarnShortSecretValue(value: string, threshold: number): boolean {
-  return isShortSecretValue(value, threshold);
-}
-
-export function shouldWarnSensitiveVariableName(key: string): boolean {
-  return isSensitiveSecretName(key);
-}
