+++ publication-closure.json
@@ -50,7 +50,6 @@
     "@shipfox/api-email-challenges",
     "@shipfox/api-integration-core",
     "@shipfox/api-integration-core-dto",
-    "@shipfox/api-integration-spi",
     "@shipfox/api-integration-gitea",
     "@shipfox/api-integration-gitea-dto",
     "@shipfox/api-integration-github",
@@ -63,6 +62,7 @@
     "@shipfox/api-integration-sentry-dto",
     "@shipfox/api-integration-slack",
     "@shipfox/api-integration-slack-dto",
+    "@shipfox/api-integration-spi",
     "@shipfox/api-integration-webhook",
     "@shipfox/api-integration-webhook-dto",
     "@shipfox/api-logs",
